Modernize emp-tool/utils: VAES/VPCLMULQDQ kernels, faster PRG/Hash, restructured tests
Performance
- aes.h: VAES+AVX-512 ParaEnc path (4-block zmm tiles via _mm512_aesenc_epi128)
  with ymm/scalar tail. ~2-4x over scalar AES-NI on Zen 4 / Sapphire Rapids /
  Alder-P; AES-NI fallback unchanged.
- f2k.h: VPCLMULQDQ-512 path for vector_inn_prdt_sum_no_red — 4-lane carryless
  multiply with deferred per-lane fold. ~4x on inner-product hot paths.
  vector_inn_prdt_sum_red now defers reduction (2x).
- f2k.h: GaloisFieldPacking::packing(block*) uses byte-shift specialization
  instead of 128 PCLMULs (~2.5x).
- hash.h: Hash::hash_once uses thread_local persistent EVP_MD_CTX, eliminating
  ~700-1000 cy of alloc/free per call (2x on small inputs).
- prg.h: random_block matches the agmpc reference — strict-aligned, direct
  write to dest, encrypt in place, 64-block chunks for L1 locality.

Cleanup / API
- Merge aes_opt.h into aes.h; delete decrypt paths, BlockVec / default_init
  allocator, bit0_mask / bit1_mask, int_to_bool / block_to_bool helpers.
- Add const to AES_KEY/userkey throughout.
- Unify all bool/bits conversions around bools_to_bits / bits_to_bools in
  block.h; bool_to_int and bool_to_block become 1-line wrappers.
- block.h: operator<< now saves/restores stream flags+fill; set_bit asserts
  index range; add xorBlocksTo_arr (in-place __restrict__ XOR).
- mitccrh.h: setS resets gid (was a latent bug); document not mixing tweak/
  gid modes.
- prg.h: PRP(const char*) now actually uses the bytes (was silently zero).
- prg.h: PRG ctor uses _mm_loadu_si128 for unaligned seed; alignas(block)
  on PRG::buffer.
- emp-tool.h: top-level static_assert(__BYTE_ORDER__ == __ORDER_LITTLE_ENDIAN__).
- constants.h: fix_key externalized (definition in emp-tool.cpp).
- ThreadPool: C++17/20 invoke_result split.

New f2k bool overloads
- vector_inn_prdt_sum_red(block*, const block*, const bool*, int): inner
  product when b ∈ {0,1}; branchless mask + XOR with 4-way unroll. No PCLMUL.
- GaloisFieldPacking::packing(block*, const bool*): trivially equals
  bools_to_bits — Σ b[i]·X^i is just LSB-first bit-packing for bool b.

Tests
- One file per utils/*.h header. Each follows a 3-section template:
  example() (readable usage), run_correctness() (against an external ground
  truth — OpenSSL for AES/SHA-256, FIPS test vectors, scalar reference for
  GF(2^128)), bench(double sec) (size sweeps with chained-dep single-shot
  ops to defeat hoisting). New test/aes.cpp, block.cpp, ccrh.cpp, crh.cpp,
  tccrh.cpp, utils.cpp; rewritten f2k.cpp, hash.cpp, mitccrh.cpp, prg.cpp,
  prp.cpp.
- Delete legacy aes_opt.cpp, to_bool.cpp; fold hash_bench / aes_bench /
  transpose_bench into their parent test files.
- CMakeLists registers utils tests in alphabetical order.
- Add CLAUDE.md documenting the test conventions.

+++ emp-tool/utils/ThreadPool.h
@@ -47,7 +47,7 @@ public:
         -> std::future<typename std::result_of<F(Args...)>::type>;
 #endif	
     ~ThreadPool();
-	int size() const;
+	size_t size() const;
 private:
     // need to keep track of threads so we can join them
     std::vector< std::thread > workers;
@@ -60,7 +60,7 @@ private:
     bool stop;
 };
  
-int inline ThreadPool::size() const {
+size_t inline ThreadPool::size() const {
 	return workers.size();
 }
 // the constructor just launches some amount of workers
